refactor: make library warning-clean under -Wall -Wextra
Fix all 67 -Wall -Wextra warnings in src/ without changing behavior, and
replace the non-standard <memory.h> with <string.h> throughout:

- sign-compare: cast non-negative int counters at unsigned comparison sites
- maybe-uninitialized: initialize locals to neutral defaults
- pointer-sign: cast unsigned char* <-> char* at the boundary (IBMP/IGIF/IXPM)
- unused-parameter: explicit (void) for unused 'options' params
- unused-result: check fgets/fread returns in IPPM, returning IFileInvalid
- absolute-value: use fabs() for doubles in IDrawLin (also fixes a latent dash
  bug where abs() truncated the slope to int)
- unused-but-set / type-limits / int-to-pointer: remove dead vars; drop the
  always-true unsigned >= 0 test; cast via intptr_t in IColor

IWriteImageFile now returns IFunctionNotImplemented for formats without a
write function instead of an uninitialized value (fixes both the warning and
a real garbage-return bug).

+++ src/ITextDim.c
@@ -19,7 +19,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <ctype.h>
-#include <memory.h>
+#include <string.h>
 
 #include "Ilib.h"
 #include "IlibP.h"
@@ -92,7 +92,7 @@ unsigned int *height_return;
 
   ret_height = font_height;
 
-  for ( ptr = text, loop = 0; loop < len; loop++, ptr++ ) {
+  for ( ptr = text, loop = 0; (unsigned int) loop < len; loop++, ptr++ ) {
     if ( *ptr == '\012' ) {
       charx = 0;
       chary += font_height;
@@ -115,7 +115,7 @@ unsigned int *height_return;
     else {
       loop2 = 0;
       ptr++;
-      while ( *ptr != ';' && loop < len && loop2 < 256 ) {
+      while ( *ptr != ';' && (unsigned int) loop < len && loop2 < 256 ) {
         ch[loop2] = *ptr;
         ptr++;
         loop++;
